16.proactive-messages/requirements.txt
botbuilder-integration-aiohttp>=4.15.0
httpx>=0.28.0
